Rid Timekit memory leak
Found the timekitDB.js file left in the db repository that is
required in the server's index.js.  This file serves no purpose
but instead takes up space on our server that could be easily
delegated to other more pressing matters. Deleted file and deleted
line of code that requires the file into index.js.